+++ src/Views/BucketPanel/EntryDetail.tsx
@@ -33,7 +33,6 @@ import "codemirror/lib/codemirror.css";
 import "codemirror/mode/javascript/javascript";
 import UploadFileForm from "../../Components/Entry/UploadFileForm";
 
-import streamSaver from "streamsaver";
 import { getExtensionFromContentType } from "../../Helpers/contentType";
 
 // @ts-ignore
@@ -263,32 +262,33 @@ export default function EntryDetail(props: Readonly<Props>) {
 
     try {
       const bucket = await props.client.getBucket(bucketName);
-      const readableRecord = await bucket.beginRead(
+      const fileName = generateFileName(record.key, record.contentType);
+      // Set expiration time for 1 hour from now
+      const expireAt = new Date(Date.now() + 60 * 60 * 1000);
+      const shareLink = await bucket.createQueryLink(
         entryName,
         BigInt(record.key),
+        undefined,
+        undefined,
+        0,
+        expireAt,
+        fileName,
       );
-      const fileName = generateFileName(record.key, record.contentType);
-      const size = Number(readableRecord.size);
-      if (size < 1024 * 1024) {
-        // Small file: use Blob and anchor
-        const data = await readableRecord.read();
-        const blob = new Blob([data as BlobPart], { type: record.contentType });
-        const url = URL.createObjectURL(blob);
-        const a = document.createElement("a");
-        a.href = url;
-        a.download = fileName;
-        a.click();
-        URL.revokeObjectURL(url);
-      } else {
-        // Large file: stream to disk
-        const fileStream = streamSaver.createWriteStream(fileName, { size });
-        await readableRecord.stream.pipeTo(fileStream);
-      }
+      const a = document.createElement("a");
+      a.href = shareLink;
+      a.download = fileName;
+      a.style.display = "none";
+      document.body.appendChild(a);
+      a.click();
+      document.body.removeChild(a);
     } catch (err) {
       console.error("Download failed", err);
       message.error("Failed to download record");
     } finally {
-      setDownloadingKey(null);
+      // cannot track download completion, wait for 2 seconds
+      setTimeout(() => {
+        setDownloadingKey(null);
+      }, 2000);
     }
   };