+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix control buttons with write permissions when token with wildcard is used, [PR-138](https://github.com/reductstore/web-console/pull/138)
+- Fixed large record download failures by using shared links, [PR-140](https://github.com/reductstore/web-console/pull/140)
 
 ## [1.11.2] - 2025-08-09